fix tests

+++ api/config.go
@@ -21,7 +21,7 @@ type Config struct {
 	GasStation      gasstation.Config `yaml:"gasStation"`
 	RangeQueryLimit uint64            `yaml:"rangeQueryLimit"`
 	Tracer          tracer.Config     `yaml:"tracer"`
-	BatchSizeLimit  uint64            `yaml:"batchSizeLimit"`
+	BatchSizeLimit  uint16            `yaml:"batchSizeLimit"`
 }
 
 // DefaultConfig is the default config
